Refactored MVC structure
