+++ pathfinding/generator/pathsGenerator.py
@@ -4,24 +4,33 @@ from models.Path import Path
 def chooseRandomCell(availableCells):
     return random.choice(availableCells) if len(availableCells) > 0 else None
 
+def chooseRandomItem(s):
+    return random.choice(list(s)) if len(s) > 0 else None
 
 def createPaths(nAgents, max, graph):
-    # for all nAgents we will choose randomly the initial and goal positions
+    # For all nAgents we will choose randomly the initial and goal positions
     # The movement of the agents will be random as well
+    # INIT and GOAL must be different for each agent
 
-    availableCells = list(graph.adjacent.keys())
-
-    #INIT and GOAL must be different for each agent
-    if len(availableCells) < nAgents*2:
+    availableCells = set(graph.adjacent.keys())
+    
+    if len(availableCells) < nAgents:
         print("Not enough cells to create a path for each agent")
         return []
     
     goals = set()
 
+    # for _ in range(nAgents):
+    #     goal = chooseRandomCell(availableCells)
+    #     while goal in goals:
+    #         goal = chooseRandomCell(availableCells)
+    #     goals.add(goal)
     for _ in range(nAgents):
-        goal = chooseRandomCell(availableCells)
+        goal = availableCells.pop()
         while goal in goals:
-            goal = chooseRandomCell(availableCells)
+            availableCells.add(goal)
+            goal = availableCells.pop()
+        availableCells.add(goal)
         goals.add(goal)
 
     # create path for each agent
@@ -29,23 +38,23 @@ def createPaths(nAgents, max, graph):
     for i in range(nAgents):
         path = Path()
 
-        # init = chooseRandomCell(availableCells)
-        # availableCells.remove(init)
-
-        # #TODO: is it right to remove init from avaibleCells? A goal could be the init of another agent
-        # goal = chooseRandomCell(availableCells)
-
         goal = goals.pop()
 
-        init = chooseRandomCell(availableCells)
-        while init[0] == goal[0] and init[1] == goal[1]:
-            print("Init and goal of agents: ", i, " are the same")
-            print(init)
-            print(goal)
-            print(availableCells)
-            print("----")
-            init = chooseRandomCell(availableCells)
-        availableCells.remove(init)
+        if goal in availableCells:
+            availableCells.remove(goal)
+            init = availableCells.pop()
+            availableCells.add(goal)
+        else:
+            init = availableCells.pop()
+        
+        # while init[0] == goal[0] and init[1] == goal[1]:
+        #     print("Init and goal of agents: ", i, " are the same")
+        #     print(init)
+        #     print(goal)
+        #     print(availableCells)
+        #     print("----")
+        #     init = chooseRandomCell(availableCells)
+        # availableCells.remove(init)
 
         current = init
         # create random path from init to goal
@@ -69,7 +78,7 @@ def createPaths(nAgents, max, graph):
             path.addMove(t, current, move.dst, move.weight)
             current = move.dst
             t += 1
-            if current == goal:
+            if current[0] == goal[0] and current[1] == goal[1]:
                 break
         
         paths.append(path)
@@ -78,13 +87,5 @@ def createPaths(nAgents, max, graph):
         path.printPath()
     
     return paths
-
-
-# nrows, ncols = 3, 3
-# freeCellRatio = 0.8
-# # these variables will be all parameters of the instance
-
-# graph = graphGenerator(nrows, ncols, freeCellRatio)
-# main(2, 20, graph)
